% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-block.R
\name{block_output.lm_block}
\alias{block_output.lm_block}
\title{Custom block output for lm_block}
\usage{
\method{block_output}{lm_block}(x, result, session)
}
\description{
Displays the model summary as text output.
}
