% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-block.R
\name{block_output.model_block}
\alias{block_output.model_block}
\title{Custom block output for model_block}
\usage{
\method{block_output}{model_block}(x, result, session)
}
\description{
Displays the model summary as text output.
}
