% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-3d-block.R
\name{block_output.regression_3d_block}
\alias{block_output.regression_3d_block}
\title{Custom block output for regression_3d_block}
\usage{
\method{block_output}{regression_3d_block}(x, result, session)
}
\description{
Custom block output for regression_3d_block
}
