% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual-explorer-block.R
\name{block_output.residual_explorer_block}
\alias{block_output.residual_explorer_block}
\title{Custom block output for residual_explorer_block}
\usage{
\method{block_output}{residual_explorer_block}(x, result, session)
}
\description{
Custom block output for residual_explorer_block
}
