% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-3d-block.R
\name{block_ui.regression_3d_block}
\alias{block_ui.regression_3d_block}
\title{Custom block UI for regression_3d_block}
\usage{
\method{block_ui}{regression_3d_block}(id, x, ...)
}
\description{
Custom block UI for regression_3d_block
}
