% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual-explorer-block.R
\name{block_ui.residual_explorer_block}
\alias{block_ui.residual_explorer_block}
\title{Custom block UI for residual_explorer_block}
\usage{
\method{block_ui}{residual_explorer_block}(id, x, ...)
}
\description{
Custom block UI for residual_explorer_block
}
