% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.lm-package}
\alias{blockr.lm}
\alias{blockr.lm-package}
\title{blockr.lm: Linear Model Blocks for blockr}
\description{
Extends 'blockr.core' with interactive blocks for statistical model fitting and diagnostics. Users can build and explore linear and generalized linear models through a graphical interface, including coefficient tables, ANOVA tables, diagnostic plots, and interactive residual exploration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blockr-org/blockr.lm}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

}
\keyword{internal}
