% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic-plot-block.R
\name{create_diagnostic_plots}
\alias{create_diagnostic_plots}
\title{Create diagnostic plots for lm object}
\usage{
create_diagnostic_plots(model)
}
\arguments{
\item{model}{An lm model object}
}
\value{
NULL (plots are created as side effect)
}
\description{
Internal helper function that creates the 4-panel diagnostic plot.
}
\keyword{internal}
