% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova-block.R
\name{new_anova_block}
\alias{new_anova_block}
\title{ANOVA Table Block}
\usage{
new_anova_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_transform_block]{blockr.core::new_transform_block()}}}
}
\value{
A transform block object of class \code{anova_block} that outputs a data frame
with columns: term, df, sumsq, meansq, statistic, p.value.
}
\description{
Generate an ANOVA table from a linear model object. Shows the decomposition
of variance for each model term with degrees of freedom, sum of squares,
mean squares, F-statistics, and p-values.
}
\examples{
if (interactive()) {
  library(blockr.core)
  model <- lm(mpg ~ cyl + hp, data = mtcars)
  serve(new_anova_block(), list(data = model))
}

}
