% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary-block.R
\name{new_model_summary_block}
\alias{new_model_summary_block}
\title{Model Summary Block}
\usage{
new_model_summary_block(...)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_transform_block]{blockr.core::new_transform_block()}}}
}
\value{
A transform block object of class \code{model_summary_block}.
}
\description{
Creates an HTML summary display for any model with broom methods.
Shows coefficients, model statistics, and diagnostic tests in a
three-column layout similar to blockr.seasonal.
}
\details{
Works with lm, glm, lmer, coxph, gam, and 100+ other model types
supported by broom.
}
\examples{
if (interactive()) {
  library(blockr.core)
  model <- lm(mpg ~ cyl + hp + wt, data = mtcars)
  serve(new_model_summary_block(), list(data = model))
}

}
